7-Zip Version 4.65

+++ CPP/Windows/Control/Trackbar.h
@@ -20,7 +20,7 @@ public:
     { SendMessage(TBM_SETTICFREQ, freq); }
   
   int GetPos()
-    { return SendMessage(TBM_GETPOS); }
+    { return (int)SendMessage(TBM_GETPOS); }
 };
 
 }}
